requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
aiosqlite==0.19.0
httpx==0.27.0
feedparser==6.0.11
pydantic==2.7.1
pydantic-settings==2.2.1
APScheduler==3.10.4
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
openai>=1.43.0
beautifulsoup4==4.12.3
yfinance==0.2.43

